+++ example/src/components/Card/Card.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 
-import {classList, StylableProps, StyledView} from "style-composer";
-import {$Card}                                from "./Card.style";
+import {StylableProps, StyledView} from "style-composer";
+import {$Card}                     from "./Card.style";
 
 export interface CardProps extends StylableProps {
   children?: React.ReactNode;
@@ -9,7 +9,7 @@ export interface CardProps extends StylableProps {
 
 const Card = (props: CardProps) => {
   const {children, classes, style} = props;
-  return <StyledView tag={"span"} style={[style]} classes={classList($Card, classes)}>
+  return <StyledView tag={"span"} style={[style]} classes={[$Card, classes]}>
     {children}
   </StyledView>;
 };